requirements.txt
langchain
pinecone
langchain-google-genai
pillow
langchain-pinecone
python-dotenv
pypdf
sentence-transformers
chainlit
-e .